[Android]Change the name of javascript object used in instrumentation tests
Rename the original js object echo to echoJava to indicate that the object
stands for extensions written in Java. Also, the corresponding html files
used for testing are changed to echo_java.html/echo_sync_java.html

The purpose of renaming is to avoid conflicting if the extension is written
in C++. We need a unique name to differentiate these objects.